refactor(ui): remove ModeSelector component
- Deleted the ModeSelector component that was used for switching between activity modes
- Hardcoded application to use 'my-activity' mode
- Moved ActivityMode type to common.ts
- Updated all imports and dependencies
- Updated tests to match new implementation
